requirements.txt
absl-py==2.0.0
accelerate==0.23.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
attrs==23.1.0
av==11.0.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.28.4
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
#deepspeed==0.12.6
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
ffmpy==0.3.2
filelock==3.12.4
#flash-attn==2.5.6
fonttools==4.48.1
fsspec==2023.9.2
ftfy==6.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
imutils==0.5.4
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
katna==0.9.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==18.1.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
#nvidia-cublas-cu11==11.10.3.66
#nvidia-cuda-cupti-cu11==11.7.101
#nvidia-cuda-nvrtc-cu11==11.7.99
#nvidia-cuda-runtime-cu11==11.7.99
#nvidia-cudnn-cu11==8.5.0.96
#nvidia-cufft-cu11==10.9.0.58
#nvidia-curand-cu11==10.2.10.91
#nvidia-cusolver-cu11==11.4.0.1
#nvidia-cusparse-cu11==11.7.4.91
#nvidia-ml-py==12.535.133
#nvidia-nccl-cu11==2.14.3
#nvidia-nvtx-cu11==11.7.91
nvitop==1.3.2
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.13.3
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.8.1.78
pandas==2.0.3
parameterized==0.9.0
pathtools==0.1.2
peft==0.5.0
Pillow==10.0.1
portalocker==2.8.2
proglog==0.1.10
protobuf==4.24.4
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==1.10.13
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rouge==1.0.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
shortuuid==1.0.11
smmap==5.0.1
sniffio==1.3.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.14.1
#torch==2.0.1
#torchaudio==2.0.2
#torchvision==0.15.2
tqdm==4.66.1
transformers==4.34.0
triton==2.0.0
tzdata==2024.1
urllib3==2.0.6
webdataset==0.2.86
Werkzeug==3.0.0
yacs==0.1.8
zipp==3.17.0
